♻️ extract add column button component
